cleanup

+++ testing/trino-testing/src/main/java/io/trino/testing/sql/TestTable.java
@@ -35,29 +35,34 @@ public class TestTable
 
     private final SqlExecutor sqlExecutor;
     private final String name;
+    private final Boolean supportSqlCreateAndDrop;
 
-    public TestTable(SqlExecutor sqlExecutor, String namePrefix, String tableDefinition) {
+    public TestTable(SqlExecutor sqlExecutor, String namePrefix, String tableDefinition)
+    {
         this(sqlExecutor, namePrefix, tableDefinition, true);
     }
 
-    public TestTable(SqlExecutor sqlExecutor, String namePrefix, Boolean supportInsertDataByJdbc) {
-        this(sqlExecutor, namePrefix, "dummy table definition", ImmutableList.of(), supportInsertDataByJdbc);
-    }
-
-    public TestTable(SqlExecutor sqlExecutor, String namePrefix, String tableDefinition, Boolean supportInsertDataByJdbc)
+    public TestTable(SqlExecutor sqlExecutor, String namePrefix, Boolean supportSqlCreateAndDrop)
     {
-        this(sqlExecutor, namePrefix, tableDefinition, ImmutableList.of(), supportInsertDataByJdbc);
+        this(sqlExecutor, namePrefix, "dummy table definition", ImmutableList.of(), supportSqlCreateAndDrop);
     }
 
-    public TestTable(SqlExecutor sqlExecutor, String namePrefix, String tableDefinition, List<String> rowsToInsert) {
+    public TestTable(SqlExecutor sqlExecutor, String namePrefix, String tableDefinition, Boolean supportSqlCreateAndDrop)
+    {
+        this(sqlExecutor, namePrefix, tableDefinition, ImmutableList.of(), supportSqlCreateAndDrop);
+    }
+
+    public TestTable(SqlExecutor sqlExecutor, String namePrefix, String tableDefinition, List<String> rowsToInsert)
+    {
         this(sqlExecutor, namePrefix, tableDefinition, rowsToInsert, true);
     }
 
-    public TestTable(SqlExecutor sqlExecutor, String namePrefix, String tableDefinition, List<String> rowsToInsert, Boolean supportInsertDataByJdbc)
+    public TestTable(SqlExecutor sqlExecutor, String namePrefix, String tableDefinition, List<String> rowsToInsert, Boolean supportSqlCreateAndDrop)
     {
+        this.supportSqlCreateAndDrop = supportSqlCreateAndDrop;
         this.sqlExecutor = sqlExecutor;
         this.name = namePrefix + "_" + randomTableSuffix();
-        if(supportInsertDataByJdbc) {
+        if (supportSqlCreateAndDrop) {
             sqlExecutor.execute(format("CREATE TABLE %s %s", name, tableDefinition));
             try {
                 for (String row : rowsToInsert) {
@@ -71,11 +76,6 @@ public class TestTable
                 }
             }
         }
-    }
-
-    public String getName()
-    {
-        return name;
     }
 
     public static TestTable fromColumns(SqlExecutor sqlExecutor, String namePrefix, Map<String, List<String>> columns)
@@ -129,15 +129,22 @@ public class TestTable
         return new TestTable(sqlExecutor, namePrefix, tableDefinition, rows.build());
     }
 
-    @Override
-    public void close()
-    {
-        sqlExecutor.execute("DROP TABLE " + name);
-    }
-
     public static String randomTableSuffix()
     {
         String randomSuffix = Long.toString(abs(random.nextLong()), MAX_RADIX);
         return randomSuffix.substring(0, min(RANDOM_SUFFIX_LENGTH, randomSuffix.length()));
+    }
+
+    public String getName()
+    {
+        return name;
+    }
+
+    @Override
+    public void close()
+    {
+        if (supportSqlCreateAndDrop) {
+            sqlExecutor.execute("DROP TABLE " + name);
+        }
     }
 }
